Related to #640
