+++ coderdojochi/views/profile.py
@@ -22,10 +22,11 @@ from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import TemplateView
 
 import arrow
+from accounts.forms import GuardianForm, MentorForm
 from icalendar import Calendar, Event, vText
 from paypal.standard.forms import PayPalPaymentsForm
 
-from coderdojochi.forms import CDCModelForm, ContactForm, DonationForm, GuardianForm, MentorForm, StudentForm
+from coderdojochi.forms import CDCModelForm, ContactForm, DonationForm, StudentForm
 from coderdojochi.models import (
     Donation,
     Equipment,
