+++ coderdojochi/views/sessions.py
@@ -23,11 +23,12 @@ from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import RedirectView, TemplateView, View
 
 import arrow
+from accounts.forms import GuardianForm, MentorForm
 from dateutil.relativedelta import relativedelta
 from icalendar import Calendar, Event, vText
 from paypal.standard.forms import PayPalPaymentsForm
 
-from coderdojochi.forms import CDCModelForm, ContactForm, DonationForm, GuardianForm, MentorForm, StudentForm
+from coderdojochi.forms import CDCModelForm, ContactForm, DonationForm, StudentForm
 from coderdojochi.mixins import RoleRedirectMixin
 from coderdojochi.models import (
     Donation,
